Derive ranges for the internal nodes of a tree

Derive ranges for the internal nodes of a tree

Usage

build_tree_ranges(tree, tip_occs)

Arguments

tree

Phylogeny (object of class "phylo").

tip_occs

Site-by-taxon matrix of occurrences for terminal taxa.

Value

A site-by-taxon matrix with a column for every branch, including terminals and internal edges.