Null model randomization analysis of alpha diversity metrics

This function compares to diversity metrics calculated in sphy_diversity to their null distributions computed by randomizing the community matrix. Randomization is done using the strand method for community matrices containing continuous occurrence quantities such as occurrence probabilities or abundances.

Usage

sphy_rand(sp, n_rand = 100, spatial = T, n_cores = 1, ...)

Arguments

sp

spatialphy object.

n_rand

Integer giving the number of random communities to generate.

spatial

Boolean: should the function return a spatial object (TRUE, default) or a matix (FALSE).

n_cores

Integer giving the number of compute cores to use for parallel processing.

...

Additional arguments passed to strand, such as n_strata, jitter, transform, priority, etc.

Value

A matrix with a row for every row of x, a column for every metric in sphy_diversity, and values indicating the proportion of randomizations in which the observed diversity metric was greater than the randomized metric.