Stratified randomization of community matrix

This is a community null model method for quantitative data (e.g. abundance or occurrence probability) that preserves row and column totals, and also approximately preserves the marginal distributions of rows and columns. For each randomization the data set is split into strata, each representing a numerical range of the input quantities, a separate binary randomization is done for each stratum, and the results are combined to produce a randomized, quantitative community matrix. See vegan::commsim() for details about other binary and quantitative null models.

Usage

strand(
  x,
  n_strata = 5,
  transform = identity,
  algorithm = "curveball",
  jitter = 0.99,
  priority = "neither",
  ...
)

Arguments

x

Community matrix with species in rows, sites in columns, and nonnegative occurrence quantities in cells.

n_strata

Integer giving the number of strata to split the data into. Must be 2 or greater. Larger values will result randomizations with less mixing but higher fidelity to marginal distributions.

transform

A function used to transform the values in x before assigning them to n_strata equal intervals. Examples include sqrt, log, rank, etc.; the default is identity.

algorithm

Null model algorithm, passed to vegan::nullmodel. Testing has only been done with the "curveball" algorithm, so other options should be use with caution. Only binary methods should be used.

jitter

Number between 0 and 1, indicating how much to randomly jitter the location of stratum boundaries.

priority

Either "rows", "cols", or "neither", indicating whether randomization within strata should prioritize maintaining the marginal distributions of the rows or columns of the input matrix. The default, "neither", prioritizes both.

n_iter

Positive integer giving the number of iterations to use for stepwise algorithms.

Value

A randomized version of x.